Null dereference in SVGTextElement::animatedLocalTransform() https://bugs.webkit.org/show_bug.cgi?id=82375
Reviewed by Nikolas Zimmermann.

Source/WebCore:

A poorly structured SVG file with a text element inside an
animatedTransform element leaves the text element placed outside of
the SVG root and it has no renderer(). Later attempts to use the
renderer in animatedLocalTransform fail.

Test: svg/custom/get-text-element-transform-crash.html

* svg/SVGTextElement.cpp:
(WebCore::SVGTextElement::animatedLocalTransform): Added a check for null renderer, and do not check
for style transform if there is no renderer.

LayoutTests:

* svg/custom/get-text-element-transform-crash-expected.txt: Added.
* svg/custom/get-text-element-transform-crash.html: Added.


git-svn-id: svn://svn.chromium.org/blink/trunk@112394 bbb929c8-8fbe-4397-9dbb-9b2b20218538