Break the association between form controls and their owner when the owner leaves the tree https://bugs.webkit.org/show_bug.cgi?id=81420
Reviewed by Kent Tamura.

Source/WebCore:

When parsing malformed markup, it's possible for a form and a
form-associated element to not be in an ancestor/descendant relationship.
When that form is later removed from the tree, the relationship needs to be updated.

Test: fast/forms/parser-associated-form-removal.html

* html/FormAssociatedElement.cpp:
(WebCore::FormAssociatedElement::FormAssociatedElement):
(WebCore::FormAssociatedElement::formRemovedFromTree): New method called when
by the form when it's being removed: handles either resetting the
owner or storing a bit to optimize away a second walk up the tree.
* html/FormAssociatedElement.h:
(FormAssociatedElement):
* html/HTMLFormElement.cpp:
(WebCore::findRoot): Simple helper method copied from FormAssociatedElement (this should probably live elsewhere).
(WebCore::HTMLFormElement::removedFromTree): Override removedFromTree to notify associated elements of the removal.
* html/HTMLFormElement.h:
(HTMLFormElement):

LayoutTests:

* fast/forms/parser-associated-form-removal-expected.txt: Added.
* fast/forms/parser-associated-form-removal.html: Added.


git-svn-id: svn://svn.chromium.org/blink/trunk@113817 bbb929c8-8fbe-4397-9dbb-9b2b20218538