REGRESSION (r94497): Pressing Command+A when inline (Marked Text) is not empty will clean whole content https://bugs.webkit.org/show_bug.cgi?id=84501
Reviewed by Alexey Proskuryakov.

Source/WebCore: 

The bug was caused by setComposition, which is called by cancelComposition, deleting the contents when
the passed text is empty. Fixed it by not deleting text when canceling compositions. This is okay because
as the comment above the line suggests, this particular call to TypingCommand::deleteSelection is only useful
when the confirmed text is empty and the composition text had previously been non-empty.

Test: editing/input/select-all-clear-input-method.html

* editing/Editor.cpp:
(WebCore::Editor::setComposition):

LayoutTests: 

Add a regression test for selecting all when there is a composition text.

* editing/input/select-all-clear-input-method-expected.txt: Added.
* editing/input/select-all-clear-input-method.html: Added.


git-svn-id: svn://svn.chromium.org/blink/trunk@115409 bbb929c8-8fbe-4397-9dbb-9b2b20218538