Inline Node::traverseNextNode https://bugs.webkit.org/show_bug.cgi?id=85844
Reviewed by Ryosuke Niwa.

Inline traverseNextNode and traverseNextSibling to reduce entry/exit overhead and allow better code generation
for many hot loops. Also added separate versions of stayWithin and unscoped cases (the function is
so simple that this seemed like the cleanest way to do it, the most reliable too) and used UNLIKELY for the 
end-of-traversal conditions.
        
The traversal function can show up to ~1% in normal page loading profiles.
        
run-perf-tests seems to think this is a progression in some subtests though bots will tell for certain.

* WebCore.exp.in:
* dom/ContainerNode.h:
        
    Following the existing pattern, function bodies go to ContainerNode.h so they can call parentNode().
    (which returns ContainerNode, not Node).

(WebCore::Node::traverseNextNode):
(WebCore):
(WebCore::Node::traverseNextSibling):
* dom/Node.cpp:
(WebCore):
* dom/Node.h:
(Node):



git-svn-id: svn://svn.chromium.org/blink/trunk@116402 bbb929c8-8fbe-4397-9dbb-9b2b20218538