WebKit erroneously add 1px padding in input elements https://bugs.webkit.org/show_bug.cgi?id=86784
Reviewed by Kent Tamura.

Source/WebCore: 

Fixed the bug by removing the code that added left and right paddings in input elements.
The code was added in r13567 to match a very old version of Internet Explorer but this quirk
does not seem to exists in IE9 (both in quirks and standard modes).

A lot of tests need to be rebaselined due to 1px difference inside input elements, and 2px
differences for the width of input elements. (In some extreme cases, line breaking changed
due to the width change).

Test: fast/forms/input-textarea-padding-match.html

* css/html.css:
(input::-webkit-input-placeholder, isindex::-webkit-input-placeholder):
* rendering/RenderTextControlSingleLine.cpp:
(WebCore::RenderTextControlSingleLine::createInnerTextStyle):

LayoutTests: 

Add a regression test.

Rebaseline lots of tests per getting rid of extra 1px padding on left and right will be done
in separate commits. Input elements' widths got shrunk by 2px in most cases and the text moved
1px to the start (left for LTR and right for RTL).

* fast/forms/input-textarea-padding-match-expected.html: Added.
* fast/forms/input-textarea-padding-match.html: Added.



git-svn-id: svn://svn.chromium.org/blink/trunk@117672 bbb929c8-8fbe-4397-9dbb-9b2b20218538