[chromium] Layers on main thread should get a RenderSurface for animating transform only if masksToBounds is true also https://bugs.webkit.org/show_bug.cgi?id=87184
Reviewed by James Robinson.

Source/WebCore:

Layers on main thread get a RenderSurface if their impl counterpart may
have a RenderSurface due to its animating transform. We currently create
a RenderSurface for all layers with an animating transform and
a descendant that drawsContents. But a RenderSurface on impl also
implies that the layer masksToBounds. So only make a RenderSurface on
main thread when the layer also masksToBounds.

Modified unit test CCLayerTreeHostCommonTest.verifyAnimationsForRenderSurfaceHierarchy.

* platform/graphics/chromium/cc/CCLayerTreeHostCommon.cpp:
(WebCore::subtreeShouldRenderToSeparateSurface):

Source/WebKit/chromium:

* tests/CCLayerTreeHostCommonTest.cpp:
(WebKitTests::TEST):


git-svn-id: svn://svn.chromium.org/blink/trunk@118219 bbb929c8-8fbe-4397-9dbb-9b2b20218538