REGRESSION: compositing/video/video-poster.html fails on Mac https://bugs.webkit.org/show_bug.cgi?id=87199
Reviewed by Maciej Stachowiak.

No new tests; fixes failing compositing/video/video-poster.html test.

Instead of creating the video layer directly, simply allow the layer
to be created in updateStates() by changing the definition of
isReadyForVideoSetup() to bypass the m_isAllowedToRender check if
the player reports a video track is present.  This causes the video layer
to be created and for future calls to prepareForRendering() to result
in calls to mediaPlayerRenderingModeChanged().

* platform/graphics/avfoundation/MediaPlayerPrivateAVFoundation.cpp:
(WebCore::MediaPlayerPrivateAVFoundation::isReadyForVideoSetup):
(WebCore::MediaPlayerPrivateAVFoundation::metadataLoaded):


git-svn-id: svn://svn.chromium.org/blink/trunk@118314 bbb929c8-8fbe-4397-9dbb-9b2b20218538